Adding Trailing Effect
A .mouser-over CSS class was added to the CSS file changing the background-color of the target
element to yellow.

In JS file, a event listener was added at the div level, the one that contains all the pixel divs.
Using bubbling effect, we check which target has the 'pixel-div' and then we change its background
color.