Dynamic Dimensions
A button was added at the top of the grid container.

I removed some CSS rules that made adding all the others easier and yet I left
the red border on the container that contains the divs.

A click event listener was added to the reset button.
On click the user is prompt to enter how many divs the container will include per side.
In order to achieve this, the CSS rule applied to div with class "div-container" is calculated based
on the default dimensions of each div (16x16) using template string.

Having determined the "max-width", the for loops append the divs to the "father" div based on user's input.